+++ R/set_na.R
@@ -106,7 +106,6 @@
 #'
 #' @export
 set_na <- function(x, ..., na, drop.levels = TRUE, as.tag = FALSE, value) {
-
   # check deprecated arguments
   if (!missing(value)) {
     message("Argument `value` is deprecated. Please use `na` instead.")
@@ -145,6 +144,7 @@ set_na <- function(x, ..., na, drop.levels = TRUE, as.tag = FALSE, value) {
 
 #' @importFrom stats na.omit
 #' @importFrom haven tagged_na na_tag
+#' @importFrom sjlabelled get_values get_labels remove_labels
 set_na_helper <- function(x, value, drop.levels, as.tag) {
   # check if values has only NA's
   if (sum(is.na(x)) == length(x)) return(x)
@@ -155,14 +155,14 @@ set_na_helper <- function(x, value, drop.levels, as.tag) {
   # check if value is a named vector
   na.names <- names(value)
   # get values for value labels
-  lab.values <- get_values(x, drop.na = F)
+  lab.values <- sjlabelled::get_values(x, drop.na = F)
 
   # no tagged NA's for date values
   if (inherits(x, "Date")) as.tag <- F
 
   # get value labels
   val.lab <-
-    get_labels(
+    sjlabelled::get_labels(
       x,
       attr.only = TRUE,
       include.values = "n",
@@ -182,7 +182,7 @@ set_na_helper <- function(x, value, drop.levels, as.tag) {
       # if we have no NA, coercing to numeric worked. Now get these
       # NA values and remove value labels from vector
       if (!anyNA(na.values)) {
-        x <- suppressWarnings(remove_labels(x, value = value))
+        x <- suppressWarnings(sjlabelled::remove_labels(x, value = value))
         value <- na.values
       }
     }
@@ -232,7 +232,8 @@ set_na_helper <- function(x, value, drop.levels, as.tag) {
   }
 
   # remove unused value labels
-  removers <- which(get_values(x) %in% value)
+  removers <- which(sjlabelled::get_values(x) %in% value)
+
   if (!is.null(removers) && !sjmisc::is_empty(removers, first.only = T)) {
     vl <- as.numeric(names(val.lab))
     names(vl) <- unname(val.lab)
@@ -254,5 +255,5 @@ set_na_helper <- function(x, value, drop.levels, as.tag) {
     attr(x, "label") <- keep.var
   }
 
-  return(x)
+  x
 }