+++ R/str_start_end.R
@@ -1,0 +1,63 @@
+#' @title Find start and end index of pattern in string
+#' @name str_start
+#' @description Find
+#'
+#' @param x A character vector.
+#' @param pattern Character string to be matched in \code{x}. \code{pattern} might also
+#'          be a regular-expression object, as returned by \code{\link[stringr]{regex}},
+#'          or any of \pkg{stringr}'s supported \code{\link[stringr]{modifiers}}.
+#'
+#' @inheritParams find_var
+#'
+#' @return A numeric vector with index of start/end position(s) of \code{pattern}
+#'          found in \code{x}, or an empty vector, if \code{pattern} was not found
+#'          in \code{x}.
+#'
+#' @examples
+#' path <- "this/is/my/fileofinterest.csv"
+#' str_start(path, "/")
+#'
+#' path <- "this//is//my//fileofinterest.csv"
+#' str_start(path, "//")
+#' str_end(path, "//")
+#'
+#' x <- c("my_friend_likes me", "your_friend likes_you")
+#' str_start(x, "_")
+#'
+#' @importFrom stringr str_locate_all coll
+#' @importFrom dplyr pull
+#' @importFrom tibble as_tibble
+#' @importFrom purrr map
+#' @export
+str_start <- function(x, pattern, ignore.case = TRUE) {
+  str_start_end(x, pattern, ignore.case, index = 1)
+}
+
+
+#' @rdname str_start
+#' @export
+str_end <- function(x, pattern, ignore.case = TRUE) {
+  str_start_end(x, pattern, ignore.case, index = -1)
+}
+
+
+str_start_end <- function(x, pattern, ignore.case, index) {
+  # get all locations of pattern
+  if (inherits(pattern, "regex"))
+    pos <- stringr::str_locate_all(x, pattern)
+  else
+    pos <- stringr::str_locate_all(x, stringr::coll(pattern, ignore_case = ignore.case))
+
+  # return starting indices
+  if (length(pos) > 1) {
+    purrr::map(pos, function(st) {
+      st %>%
+        tibble::as_tibble() %>%
+        dplyr::pull(index)
+    })
+  } else {
+    pos[[1]] %>%
+      tibble::as_tibble() %>%
+      dplyr::pull(index)
+  }
+}